refactor: split large blob modules into separate test files
This commit separates inline test code from core logic in several overgrown
"Blob" modules, placing them into dedicated `tests.rs` files. This aligns
with the Atlas persona's architectural directives to maintain clean,
navigable boundaries and avoid 2,000+ line monoliths.

Changes:
- `src/query/parser.rs` -> `src/query/parser/mod.rs` & `tests.rs`
- `src/query/converter.rs` -> `src/query/converter/mod.rs` & `tests.rs`
- `src/storage/wal/segment_reader.rs` -> `src/storage/wal/segment_reader/mod.rs` & `tests.rs`
- `src/storage/current/mod.rs` (extracted `coverage_tests` into `tests.rs`)
