+++ src/storage/current/tests.rs
@@ -1950,3 +1950,25 @@ fn test_get_node_ids_by_label() {
     let nonexistent_ids = storage.get_node_ids_by_label("Nonexistent");
     assert_eq!(nonexistent_ids.len(), 0);
 }
+
+// Migrated coverage_tests
+mod coverage_tests {
+    use super::*;
+
+    #[test]
+    fn test_vector_search_index_not_found() {
+        let storage = CurrentStorage::new();
+        // Don't enable vector index
+
+        let node_id = NodeId::new(1).unwrap();
+        // Try search
+        let result = storage.find_similar(node_id, 5);
+        assert!(result.is_err());
+        assert!(format!("{}", result.unwrap_err()).contains("Vector index is not enabled"));
+
+        // Try search with specific property
+        let result = storage.find_similar_in("embedding", node_id, 5);
+        assert!(result.is_err());
+        assert!(format!("{}", result.unwrap_err()).contains("Vector index not found"));
+    }
+}
