--- src/storage/wal/segment_reader.rs
@@ -1,2232 +1,0 @@
-//! WAL Segment Reader.
-//!
-//! This module provides standalone functions for reading WAL segments from disk
-//! for recovery purposes. It does not require any WAL writer state.
-//!
-//! # Memory Efficiency
-//!
-//! This module uses memory-mapped I/O (`memmap2`) to read WAL segments efficiently.
-//! Instead of loading entire segment files (default 64MB) into memory, memory-mapped
-//! files allow the OS to handle paging automatically. This provides several benefits:
-//!
-//! - **Lower memory usage**: OS pages in data as needed, not all at once
-//! - **Better caching**: OS can cache frequently accessed pages
-//! - **Automatic eviction**: OS evicts pages under memory pressure
-//! - **Reduced recovery memory**: With 10 segments, peak memory drops from 640MB+ to O(working set)
-//!
-//! See issue #216 for details.
-
-use std::fs::File;
-use std::path::Path;
-use std::sync::Arc;
-
-use crate::core::error::{Error, Result, StorageError};
-use crate::core::hlc::HybridTimestamp;
-use crate::core::id::{EdgeId, NodeId, VersionId};
-use crate::core::property::PropertyMap;
-
-use super::serialization::{
-    OP_CHECKPOINT, OP_CREATE_EDGE, OP_CREATE_NODE, OP_DELETE_EDGE, OP_DELETE_NODE, OP_UPDATE_EDGE,
-    OP_UPDATE_NODE,
-};
-use super::{LSN, WalEntry, WalOperation};
-
-/// Magic bytes identifying a AletheiaDB WAL segment file.
-pub(crate) const WAL_MAGIC: [u8; 4] = *b"GWAL";
-
-/// Current WAL format version (plaintext entries).
-pub(crate) const WAL_VERSION: u8 = 1;
-
-/// WAL format version for encrypted segments.
-///
-/// Version 2 segments use length-prefixed encrypted entries:
-/// `[4-byte LE entry length][encrypted entry bytes]`
-/// The header (magic + version) remains plaintext.
-pub(crate) const WAL_VERSION_ENCRYPTED: u8 = 2;
-
-/// Maximum supported WAL version (inclusive).
-const WAL_VERSION_MAX: u8 = WAL_VERSION_ENCRYPTED;
-
-/// Size of the WAL segment header (magic + version).
-pub(crate) const WAL_HEADER_SIZE: usize = 5;
-
-/// Maximum reasonable segment size (configurable, but 1GB is a safe upper bound)
-/// Default segments are 64MB, so 1GB allows for 16x growth
-pub(crate) const MAX_SEGMENT_SIZE: u64 = 1024 * 1024 * 1024; // 1GB
-
-/// Read all WAL entries from a directory, starting from the specified LSN.
-///
-/// This function scans the directory for segment files (*.log), reads them in order,
-/// and returns all entries with LSN >= start_lsn.
-///
-/// # Arguments
-///
-/// * `wal_dir` - Path to the WAL directory containing segment files
-/// * `start_lsn` - Only entries with LSN >= this value are returned
-///
-/// # Returns
-///
-/// A vector of WAL entries sorted by LSN.
-///
-/// # Examples
-///
-/// ```
-/// use aletheiadb::storage::wal::LSN;
-/// use aletheiadb::storage::wal::segment_reader::read_entries_from_dir;
-/// use std::fs::File;
-/// use std::io::Write;
-/// use tempfile::tempdir;
-///
-/// let dir = tempdir().unwrap();
-/// let wal_dir = dir.path();
-///
-/// // Create a dummy segment file
-/// let segment_path = wal_dir.join("0.log");
-/// let mut file = File::create(&segment_path).unwrap();
-///
-/// // Write WAL header: Magic bytes *b"GWAL" and version 1
-/// file.write_all(b"GWAL").unwrap();
-/// file.write_all(&[1]).unwrap();
-/// file.sync_all().unwrap();
-///
-/// // Read entries (returns an empty vector since we wrote no entries)
-/// let entries = read_entries_from_dir(wal_dir, LSN(1)).unwrap();
-/// assert!(entries.is_empty());
-/// ```
-pub fn read_entries_from_dir(wal_dir: &Path, start_lsn: LSN) -> Result<Vec<WalEntry>> {
-    read_entries_from_dir_with_cipher(wal_dir, start_lsn, None)
-}
-
-/// Read all WAL entries from a directory with optional decryption.
-///
-/// This function scans the directory for segment files (*.log), reads them in order,
-/// and returns all entries with LSN >= start_lsn. If a cipher is provided, version 2
-/// (encrypted) segments are decrypted transparently. Version 1 (plaintext) segments
-/// are always read without decryption regardless of the cipher parameter.
-///
-/// # Arguments
-///
-/// * `wal_dir` - Path to the WAL directory containing segment files
-/// * `start_lsn` - Only entries with LSN >= this value are returned
-/// * `cipher` - Optional cipher for decrypting version 2 segments
-///
-/// # Returns
-///
-/// A vector of WAL entries sorted by LSN.
-pub fn read_entries_from_dir_with_cipher(
-    wal_dir: &Path,
-    start_lsn: LSN,
-    cipher: Option<&Arc<dyn crate::encryption::cipher::Cipher>>,
-) -> Result<Vec<WalEntry>> {
-    let mut entries = Vec::new();
-
-    // Find all WAL segments
-    let mut segments = Vec::new();
-    if let Ok(dir_entries) = std::fs::read_dir(wal_dir) {
-        for entry in dir_entries.flatten() {
-            if let Some(name) = entry.file_name().to_str()
-                && name.ends_with(".log")
-                && let Some(seg_id) = name
-                    .strip_suffix(".log")
-                    .and_then(|s| s.parse::<u64>().ok())
-            {
-                segments.push((seg_id, entry.path()));
-            }
-        }
-    }
-
-    // Sort segments by ID
-    segments.sort_by_key(|(id, _)| *id);
-
-    // Read entries from each segment
-    for (_, path) in segments {
-        let segment_entries = read_segment_with_cipher(&path, start_lsn, cipher)?;
-        entries.extend(segment_entries);
-    }
-
-    // Sort entries by LSN to ensure correct ordering across segments.
-    // In a striped WAL architecture, entries can be flushed to different segments
-    // in an order that differs from their LSN assignment order.
-    entries.sort_by_key(|entry| entry.lsn);
-
-    Ok(entries)
-}
-
-/// Read WAL entries from a single segment file.
-///
-/// This function uses memory-mapped I/O for efficient reading without loading
-/// the entire file into memory. The OS handles paging automatically, which is
-/// especially important for large segment files (default 64MB).
-///
-/// # Arguments
-///
-/// * `path` - Path to the segment file
-/// * `start_lsn` - Only entries with LSN >= this value are returned
-///
-/// # Returns
-///
-/// A vector of WAL entries from this segment.
-///
-/// # Memory Efficiency
-///
-/// Uses `memmap2` for memory-mapped I/O. Peak memory usage is O(working set)
-/// rather than O(file size). See issue #216.
-///
-/// # Examples
-///
-/// ```
-/// use aletheiadb::storage::wal::LSN;
-/// use aletheiadb::storage::wal::segment_reader::read_segment;
-/// use std::fs::File;
-/// use std::io::Write;
-/// use tempfile::tempdir;
-///
-/// let dir = tempdir().unwrap();
-/// let segment_path = dir.path().join("segment.log");
-///
-/// // Create a segment file with a valid header: Magic bytes *b"GWAL" and version 1
-/// let mut file = File::create(&segment_path).unwrap();
-/// file.write_all(b"GWAL").unwrap();
-/// file.write_all(&[1]).unwrap();
-/// file.sync_all().unwrap();
-///
-/// // Read the empty segment
-/// let entries = read_segment(&segment_path, LSN(1)).unwrap();
-/// assert!(entries.is_empty());
-/// ```
-pub fn read_segment(path: &Path, start_lsn: LSN) -> Result<Vec<WalEntry>> {
-    read_segment_with_cipher(path, start_lsn, None)
-}
-
-/// Read WAL entries from a single segment file with optional decryption.
-///
-/// This function uses memory-mapped I/O for efficient reading. It transparently
-/// handles both version 1 (plaintext) and version 2 (encrypted) segments:
-///
-/// - **Version 1**: Entries are parsed directly (no cipher needed).
-/// - **Version 2**: Each entry is length-prefixed (`[4-byte LE len][encrypted data]`).
-///   A cipher must be provided to decrypt version 2 segments.
-///
-/// # Arguments
-///
-/// * `path` - Path to the segment file
-/// * `start_lsn` - Only entries with LSN >= this value are returned
-/// * `cipher` - Optional cipher for decrypting version 2 segments
-///
-/// # Returns
-///
-/// A vector of WAL entries from this segment.
-///
-/// # Errors
-///
-/// Returns an error if:
-/// - The segment is version 2 but no cipher is provided
-/// - Decryption fails (wrong key, corrupted data, tampered header)
-/// - The segment format is invalid
-pub fn read_segment_with_cipher(
-    path: &Path,
-    start_lsn: LSN,
-    cipher: Option<&Arc<dyn crate::encryption::cipher::Cipher>>,
-) -> Result<Vec<WalEntry>> {
-    // Open file, only treating NotFound as "empty" - all other errors are propagated
-    let file = match File::open(path) {
-        Ok(f) => f,
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(Vec::new()),
-        Err(e) => {
-            return Err(StorageError::IoError(format!(
-                "Failed to open WAL segment {:?}: {}",
-                path, e
-            ))
-            .into());
-        }
-    };
-
-    // Validate file size before mapping to prevent DoS attacks with huge files
-    let metadata = file
-        .metadata()
-        .map_err(|e| StorageError::IoError(format!("Failed to get file metadata: {}", e)))?;
-
-    if metadata.len() > MAX_SEGMENT_SIZE {
-        return Err(StorageError::CorruptedData(format!(
-            "WAL segment too large: {} bytes (max: {} bytes)",
-            metadata.len(),
-            MAX_SEGMENT_SIZE
-        ))
-        .into());
-    }
-
-    // Handle empty files explicitly to avoid mmap failure on some platforms (e.g. macOS/Windows)
-    // A zero-byte file can occur if a crash happens immediately after segment creation.
-    if metadata.len() == 0 {
-        return Ok(Vec::new());
-    }
-
-    // Memory-map the file for efficient reading without loading entire file into memory.
-    // SAFETY: We only read from the memory map, never write. The file is opened read-only.
-    // The mapping is valid for the lifetime of this function and is automatically unmapped
-    // when dropped. We have verified the file size above to prevent out-of-bounds reads.
-    let mmap = unsafe {
-        memmap2::Mmap::map(&file).map_err(|e| {
-            StorageError::IoError(format!("Failed to memory-map WAL segment: {}", e))
-        })?
-    };
-
-    // Use the memory-mapped region as a byte slice
-    let buffer = &mmap[..];
-
-    // ⚡ Bolt Optimization: Pre-allocate vector based on buffer size.
-    // Assuming an average WAL entry is ~128 bytes, this prevents numerous
-    // heap reallocations when reading large segments.
-    let capacity_hint = (buffer.len() / 128).max(1);
-    let mut entries = Vec::with_capacity(capacity_hint);
-
-    // Detect WAL format version
-    let (version, mut offset) = if buffer.len() >= WAL_HEADER_SIZE && buffer[0..4] == WAL_MAGIC {
-        // Version 1+ format: has magic header
-        let ver = buffer[4];
-        if ver > WAL_VERSION_MAX {
-            return Err(StorageError::CorruptedData(format!(
-                "Unsupported WAL version: {} (max supported: {})",
-                ver, WAL_VERSION_MAX
-            ))
-            .into());
-        }
-        (ver, WAL_HEADER_SIZE)
-    } else if !buffer.is_empty() {
-        // Invalid format: no magic header
-        return Err(StorageError::CorruptedData(
-            "Invalid WAL segment: missing GWAL magic header".to_string(),
-        )
-        .into());
-    } else {
-        return Ok(Vec::new()); // Empty segment
-    };
-
-    // Version 2 (encrypted) segments require a cipher for decryption.
-    if version == WAL_VERSION_ENCRYPTED && cipher.is_none() {
-        return Err(StorageError::Encryption(
-            "Cannot read encrypted WAL segment (version 2) without a cipher".to_string(),
-        )
-        .into());
-    }
-
-    // Dispatch to the appropriate parsing loop based on version.
-    if version == WAL_VERSION_ENCRYPTED {
-        // Version 2: length-prefixed encrypted entries.
-        let cipher = cipher.expect("cipher presence checked above");
-        parse_encrypted_entries(buffer, &mut offset, start_lsn, cipher, path, &mut entries)?;
-    } else {
-        // Version 1: plaintext entries (original format).
-        parse_plaintext_entries(buffer, &mut offset, version, start_lsn, path, &mut entries)?;
-    }
-
-    Ok(entries)
-}
-
-/// Parse plaintext (version 1) entries from a WAL segment buffer.
-fn parse_plaintext_entries(
-    buffer: &[u8],
-    offset: &mut usize,
-    version: u8,
-    start_lsn: LSN,
-    path: &Path,
-    entries: &mut Vec<WalEntry>,
-) -> Result<()> {
-    while *offset < buffer.len() {
-        match parse_entry_at(buffer, *offset, version) {
-            Ok((entry, bytes_consumed)) => {
-                if entry.lsn >= start_lsn {
-                    entries.push(entry);
-                }
-                *offset += bytes_consumed;
-            }
-            Err(e) => {
-                // Distinguish between expected EOF truncation vs. unexpected corruption
-                if *offset + 24 > buffer.len() {
-                    #[cfg(feature = "observability")]
-                    tracing::debug!(
-                        "Partial entry at end of WAL segment {:?} (offset {}/{}), stopping read",
-                        path,
-                        offset,
-                        buffer.len()
-                    );
-                    break;
-                } else {
-                    let header_slice = &buffer[*offset..*offset + 24];
-                    if header_slice.iter().all(|&b| b == 0) {
-                        #[cfg(feature = "observability")]
-                        tracing::debug!(
-                            "Zeroed region at end of WAL segment {:?} (offset {}/{}), stopping read",
-                            path,
-                            offset,
-                            buffer.len()
-                        );
-                        break;
-                    }
-
-                    #[cfg(feature = "observability")]
-                    tracing::error!(
-                        "Failed to parse WAL entry in segment {:?} at offset {}: {}",
-                        path,
-                        offset,
-                        e
-                    );
-                    #[cfg(not(feature = "observability"))]
-                    {
-                        eprintln!(
-                            "CRITICAL: Failed to parse WAL entry in segment {:?} at offset {}: {}",
-                            path, offset, e
-                        );
-                        eprintln!("Header slice: {:?}", header_slice);
-                    }
-                    return Err(e);
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Parse encrypted (version 2) entries from a WAL segment buffer.
-///
-/// Each entry is stored as `[4-byte LE length][encrypted entry bytes]`.
-/// The encrypted entry bytes are decrypted using the provided cipher,
-/// then parsed as a normal WAL entry (version 1 format).
-fn parse_encrypted_entries(
-    buffer: &[u8],
-    offset: &mut usize,
-    start_lsn: LSN,
-    cipher: &Arc<dyn crate::encryption::cipher::Cipher>,
-    path: &Path,
-    entries: &mut Vec<WalEntry>,
-) -> Result<()> {
-    while *offset < buffer.len() {
-        // Need at least 4 bytes for the length prefix
-        if *offset + 4 > buffer.len() {
-            // Partial length prefix at EOF -- truncated write
-            #[cfg(feature = "observability")]
-            tracing::debug!(
-                "Partial length prefix at end of encrypted WAL segment {:?} (offset {}/{}), stopping read",
-                path,
-                offset,
-                buffer.len()
-            );
-            break;
-        }
-
-        // Check for zeroed length prefix (indicates end of data in pre-allocated files)
-        let len_bytes: [u8; 4] = buffer[*offset..*offset + 4]
-            .try_into()
-            .expect("slice length verified above");
-        let entry_len = u32::from_le_bytes(len_bytes) as usize;
-
-        if entry_len == 0 {
-            // Zero-length entry marks end of valid data
-            break;
-        }
-
-        *offset += 4;
-
-        // Validate entry length
-        if *offset + entry_len > buffer.len() {
-            // Truncated encrypted entry at EOF
-            #[cfg(feature = "observability")]
-            tracing::debug!(
-                "Truncated encrypted entry at end of WAL segment {:?} (offset {}, entry_len {}, buf_len {}), stopping read",
-                path,
-                offset,
-                entry_len,
-                buffer.len()
-            );
-            break;
-        }
-
-        let encrypted_entry = &buffer[*offset..*offset + entry_len];
-        *offset += entry_len;
-
-        // Decrypt the entry
-        let decrypted =
-            crate::encryption::wal_encryption::decrypt_wal_payload(encrypted_entry, cipher)
-                .map_err(|e| {
-                    Error::Storage(StorageError::Encryption(format!(
-                        "Failed to decrypt WAL entry in segment {:?}: {}",
-                        path, e
-                    )))
-                })?;
-
-        // Parse the decrypted bytes as a normal (version 1) entry
-        match parse_entry_at(&decrypted, 0, WAL_VERSION) {
-            Ok((entry, _bytes_consumed)) => {
-                if entry.lsn >= start_lsn {
-                    entries.push(entry);
-                }
-            }
-            Err(e) => {
-                #[cfg(feature = "observability")]
-                tracing::error!(
-                    "Failed to parse decrypted WAL entry in segment {:?}: {}",
-                    path,
-                    e
-                );
-                #[cfg(not(feature = "observability"))]
-                eprintln!(
-                    "CRITICAL: Failed to parse decrypted WAL entry in segment {:?}: {}",
-                    path, e
-                );
-                return Err(e);
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Parse a single WAL entry from a buffer at the specified offset.
-///
-/// This function extracts the parsing logic that was previously duplicated
-/// in multiple places (issue #218). It handles all WAL operation types and
-/// returns both the parsed entry and the number of bytes consumed.
-///
-/// # Arguments
-///
-/// * `buffer` - The buffer containing serialized WAL data
-/// * `offset` - The offset in the buffer to start parsing from
-/// * `version` - The WAL format version
-///
-/// # Returns
-///
-/// A tuple of (WalEntry, bytes_consumed) on success, or an error if:
-/// - The buffer is too small to contain a valid entry
-/// - The operation type is unknown
-/// - The data is corrupted or truncated
-///
-/// # Example
-///
-/// ```ignore
-/// let (entry, bytes_consumed) = parse_entry_at(&buffer, 0, WAL_VERSION)?;
-/// let next_offset = offset + bytes_consumed;
-/// ```
-pub(crate) fn parse_entry_at(
-    buffer: &[u8],
-    offset: usize,
-    version: u8,
-) -> Result<(WalEntry, usize)> {
-    let start_offset = offset;
-    let mut current_offset = offset;
-
-    // Helper macro for checked addition to prevent overflow panics
-    macro_rules! add_offset {
-        ($n:expr) => {
-            current_offset = current_offset.checked_add($n).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })?;
-        };
-    }
-
-    // Phase 2: Need at least 24 bytes for LSN (8) + HybridTimestamp (12) + checksum (4)
-    // Use checked arithmetic for bounds check
-    if current_offset.checked_add(24).ok_or_else(|| {
-        Error::Storage(StorageError::CorruptedData(
-            "WAL offset overflow".to_string(),
-        ))
-    })? > buffer.len()
-    {
-        return Err(StorageError::CorruptedData(
-            "Insufficient buffer size for WAL entry header".to_string(),
-        )
-        .into());
-    }
-
-    // Read LSN (8 bytes)
-    let lsn = LSN(u64::from_le_bytes(
-        buffer[current_offset..current_offset + 8]
-            .try_into()
-            .unwrap(), // Safe due to buffer length check above
-    ));
-    add_offset!(8);
-
-    // Read timestamp (12 bytes: Phase 2 HybridTimestamp)
-    let (timestamp, _) = HybridTimestamp::deserialize(&buffer[current_offset..]).map_err(|e| {
-        StorageError::CorruptedData(format!("Failed to deserialize timestamp: {}", e))
-    })?;
-    add_offset!(12);
-
-    // Read checksum (4 bytes)
-    let checksum = u32::from_le_bytes(
-        buffer[current_offset..current_offset + 4]
-            .try_into()
-            .unwrap(), // Safe due to buffer length check above
-    );
-    add_offset!(4);
-
-    // Read operation type
-    if current_offset >= buffer.len() {
-        return Err(StorageError::CorruptedData(
-            "Insufficient buffer size for operation type".to_string(),
-        )
-        .into());
-    }
-    let op_type = buffer[current_offset];
-    add_offset!(1);
-
-    // Parse operation data based on type and version
-    let operation = match op_type {
-        OP_CREATE_NODE => {
-            if current_offset.checked_add(12).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })? > buffer.len()
-            {
-                return Err(StorageError::CorruptedData(
-                    "Insufficient buffer size for CreateNode".to_string(),
-                )
-                .into());
-            }
-            let node_id = deserialize_node_id(buffer, current_offset, "CreateNode")?;
-            add_offset!(8);
-
-            // Read 4-byte InternedString ID
-            if current_offset.checked_add(4).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })? > buffer.len()
-            {
-                return Err(StorageError::CorruptedData(
-                    "Insufficient buffer size for CreateNode label".to_string(),
-                )
-                .into());
-            }
-            let label_id = u32::from_le_bytes(
-                buffer[current_offset..current_offset + 4]
-                    .try_into()
-                    .unwrap(), // Safe due to buffer length check above
-            );
-            add_offset!(4);
-
-            // Reconstruct InternedString from ID
-            // During recovery, the string should already be in the interner
-            // (either from checkpoint or previous WAL entries)
-            let label = crate::core::interning::InternedString::from_raw(label_id);
-
-            // V1+: deserialize properties and temporal
-            let (properties, valid_from) = if version >= WAL_VERSION {
-                let (props, props_len) = PropertyMap::deserialize(&buffer[current_offset..])?;
-                add_offset!(props_len);
-                let (valid_from_ts, ts_len) =
-                    HybridTimestamp::deserialize(&buffer[current_offset..])?;
-                add_offset!(ts_len);
-                (props, valid_from_ts)
-            } else {
-                (PropertyMap::new(), timestamp)
-            };
-
-            WalOperation::CreateNode {
-                node_id,
-                label,
-                properties,
-                valid_from,
-            }
-        }
-        OP_CREATE_EDGE => {
-            if current_offset.checked_add(28).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })? > buffer.len()
-            {
-                return Err(StorageError::CorruptedData(
-                    "Insufficient buffer size for CreateEdge".to_string(),
-                )
-                .into());
-            }
-            let edge_id = deserialize_edge_id(buffer, current_offset, "CreateEdge")?;
-            add_offset!(8);
-
-            let source = deserialize_node_id(buffer, current_offset, "CreateEdge source")?;
-            add_offset!(8);
-
-            let target = deserialize_node_id(buffer, current_offset, "CreateEdge target")?;
-            add_offset!(8);
-
-            // Read 4-byte InternedString ID
-            if current_offset.checked_add(4).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })? > buffer.len()
-            {
-                return Err(StorageError::CorruptedData(
-                    "Insufficient buffer size for CreateEdge label".to_string(),
-                )
-                .into());
-            }
-            let label_id = u32::from_le_bytes(
-                buffer[current_offset..current_offset + 4]
-                    .try_into()
-                    .unwrap(), // Safe due to buffer length check above
-            );
-            add_offset!(4);
-
-            // Reconstruct InternedString from ID
-            let label = crate::core::interning::InternedString::from_raw(label_id);
-
-            let (properties, valid_from) = if version >= WAL_VERSION {
-                let (props, props_len) = PropertyMap::deserialize(&buffer[current_offset..])?;
-                add_offset!(props_len);
-                let (valid_from_ts, ts_len) =
-                    HybridTimestamp::deserialize(&buffer[current_offset..])?;
-                add_offset!(ts_len);
-                (props, valid_from_ts)
-            } else {
-                (PropertyMap::new(), timestamp)
-            };
-
-            WalOperation::CreateEdge {
-                edge_id,
-                source,
-                target,
-                label,
-                properties,
-                valid_from,
-            }
-        }
-        OP_UPDATE_NODE => {
-            if current_offset.checked_add(16).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })? > buffer.len()
-            {
-                return Err(StorageError::CorruptedData(
-                    "Insufficient buffer size for UpdateNode".to_string(),
-                )
-                .into());
-            }
-            let node_id = deserialize_node_id(buffer, current_offset, "UpdateNode")?;
-            add_offset!(8);
-
-            let version_id = deserialize_version_id(buffer, current_offset, "UpdateNode")?;
-            add_offset!(8);
-
-            let (label, properties, valid_from) = if version >= WAL_VERSION {
-                // Read 4-byte InternedString ID
-                if current_offset.checked_add(4).ok_or_else(|| {
-                    Error::Storage(StorageError::CorruptedData(
-                        "WAL offset overflow".to_string(),
-                    ))
-                })? > buffer.len()
-                {
-                    return Err(StorageError::CorruptedData(
-                        "Insufficient buffer size for UpdateNode label".to_string(),
-                    )
-                    .into());
-                }
-                let label_id = u32::from_le_bytes([
-                    buffer[current_offset],
-                    buffer[current_offset + 1],
-                    buffer[current_offset + 2],
-                    buffer[current_offset + 3],
-                ]);
-                add_offset!(4);
-
-                // Reconstruct InternedString from ID
-                let lbl = crate::core::interning::InternedString::from_raw(label_id);
-
-                let (props, props_len) = PropertyMap::deserialize(&buffer[current_offset..])?;
-                add_offset!(props_len);
-                let (valid_from_ts, ts_len) =
-                    HybridTimestamp::deserialize(&buffer[current_offset..])?;
-                add_offset!(ts_len);
-                (lbl, props, valid_from_ts)
-            } else {
-                (
-                    // For old WAL format, create a dummy InternedString (this shouldn't happen in practice)
-                    crate::core::interning::InternedString::from_raw(0),
-                    PropertyMap::new(),
-                    timestamp,
-                )
-            };
-
-            WalOperation::UpdateNode {
-                node_id,
-                version_id,
-                label,
-                properties,
-                valid_from,
-            }
-        }
-        OP_UPDATE_EDGE => {
-            // V0: 16 bytes (EdgeId + VersionId)
-            // V1+: 20 bytes (EdgeId + VersionId + LabelId)
-            let required = if version >= WAL_VERSION { 20 } else { 16 };
-            if current_offset.checked_add(required).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })? > buffer.len()
-            {
-                return Err(StorageError::CorruptedData(
-                    "Insufficient buffer size for UpdateEdge".to_string(),
-                )
-                .into());
-            }
-            let edge_id = deserialize_edge_id(buffer, current_offset, "UpdateEdge")?;
-            add_offset!(8);
-
-            let version_id = deserialize_version_id(buffer, current_offset, "UpdateEdge")?;
-            add_offset!(8);
-
-            let (label, properties, valid_from) = if version >= WAL_VERSION {
-                // Read 4-byte InternedString ID
-                if current_offset.checked_add(4).ok_or_else(|| {
-                    Error::Storage(StorageError::CorruptedData(
-                        "WAL offset overflow".to_string(),
-                    ))
-                })? > buffer.len()
-                {
-                    return Err(StorageError::CorruptedData(
-                        "Insufficient buffer size for UpdateEdge label".to_string(),
-                    )
-                    .into());
-                }
-                let label_id = u32::from_le_bytes([
-                    buffer[current_offset],
-                    buffer[current_offset + 1],
-                    buffer[current_offset + 2],
-                    buffer[current_offset + 3],
-                ]);
-                add_offset!(4);
-
-                // Reconstruct InternedString from ID
-                let lbl = crate::core::interning::InternedString::from_raw(label_id);
-
-                let (props, props_len) = PropertyMap::deserialize(&buffer[current_offset..])?;
-                add_offset!(props_len);
-                let (valid_from_ts, ts_len) =
-                    HybridTimestamp::deserialize(&buffer[current_offset..])?;
-                add_offset!(ts_len);
-                (lbl, props, valid_from_ts)
-            } else {
-                (
-                    // For old WAL format, create a dummy InternedString (this shouldn't happen in practice)
-                    crate::core::interning::InternedString::from_raw(0),
-                    PropertyMap::new(),
-                    timestamp,
-                )
-            };
-
-            WalOperation::UpdateEdge {
-                edge_id,
-                version_id,
-                label,
-                properties,
-                valid_from,
-            }
-        }
-        OP_CHECKPOINT => {
-            // LSN (8 bytes) + HybridTimestamp (12 bytes) = 20 bytes
-            if current_offset.checked_add(20).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })? > buffer.len()
-            {
-                return Err(StorageError::CorruptedData(
-                    "Insufficient buffer size for Checkpoint".to_string(),
-                )
-                .into());
-            }
-            let cp_lsn = LSN(u64::from_le_bytes(
-                buffer[current_offset..current_offset + 8]
-                    .try_into()
-                    .unwrap(), // Safe due to buffer length check above
-            ));
-            add_offset!(8);
-
-            // Phase 2: Deserialize HybridTimestamp (12 bytes: 8 wallclock + 4 logical)
-            let (cp_timestamp, consumed) = HybridTimestamp::deserialize(&buffer[current_offset..])?;
-            add_offset!(consumed);
-
-            WalOperation::Checkpoint {
-                lsn: cp_lsn,
-                timestamp: cp_timestamp,
-            }
-        }
-        OP_DELETE_NODE => {
-            if current_offset.checked_add(8).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })? > buffer.len()
-            {
-                return Err(StorageError::CorruptedData(
-                    "Insufficient buffer size for DeleteNode".to_string(),
-                )
-                .into());
-            }
-            let node_id = deserialize_node_id(buffer, current_offset, "DeleteNode")?;
-            add_offset!(8);
-
-            let valid_from = if version >= WAL_VERSION {
-                let (valid_from_ts, ts_len) =
-                    HybridTimestamp::deserialize(&buffer[current_offset..])?;
-                add_offset!(ts_len);
-                valid_from_ts
-            } else {
-                timestamp
-            };
-
-            WalOperation::DeleteNode {
-                node_id,
-                valid_from,
-            }
-        }
-        OP_DELETE_EDGE => {
-            if current_offset.checked_add(8).ok_or_else(|| {
-                Error::Storage(StorageError::CorruptedData(
-                    "WAL offset overflow".to_string(),
-                ))
-            })? > buffer.len()
-            {
-                return Err(StorageError::CorruptedData(
-                    "Insufficient buffer size for DeleteEdge".to_string(),
-                )
-                .into());
-            }
-            let edge_id = deserialize_edge_id(buffer, current_offset, "DeleteEdge")?;
-            add_offset!(8);
-
-            let valid_from = if version >= WAL_VERSION {
-                let (valid_from_ts, ts_len) =
-                    HybridTimestamp::deserialize(&buffer[current_offset..])?;
-                add_offset!(ts_len);
-                valid_from_ts
-            } else {
-                timestamp
-            };
-
-            WalOperation::DeleteEdge {
-                edge_id,
-                valid_from,
-            }
-        }
-        _ => {
-            // Unknown operation type
-            return Err(StorageError::CorruptedData(format!(
-                "Unknown WAL operation type: {}",
-                op_type
-            ))
-            .into());
-        }
-    };
-
-    // Verify checksum to ensure data integrity (critical for WAL correctness)
-    let mut hasher = crc32fast::Hasher::new();
-    // Hash LSN (8 bytes) + timestamp (12 bytes) = bytes 0..20
-    hasher.update(&buffer[start_offset..start_offset + 20]);
-    // Hash operation data (from after checksum field to end of entry)
-    hasher.update(&buffer[start_offset + 24..current_offset]);
-    let computed_checksum = hasher.finalize();
-
-    if checksum != computed_checksum {
-        return Err(StorageError::CorruptedData(format!(
-            "WAL entry checksum mismatch for LSN {}: expected {:#x}, got {:#x}. Entry is corrupted.",
-            lsn.0, checksum, computed_checksum
-        ))
-        .into());
-    }
-
-    let entry = WalEntry {
-        lsn,
-        timestamp,
-        operation,
-        checksum,
-    };
-
-    let bytes_consumed = current_offset - start_offset;
-    Ok((entry, bytes_consumed))
-}
-
-/// Helper to deserialize and validate a NodeId from WAL buffer
-#[inline]
-fn deserialize_node_id(buffer: &[u8], offset: usize, context: &str) -> Result<NodeId> {
-    let bytes = buffer.get(offset..offset + 8).ok_or_else(|| {
-        Error::Storage(StorageError::CorruptedData(format!(
-            "Insufficient buffer size for NodeId in {}",
-            context
-        )))
-    })?;
-    let raw_id = u64::from_le_bytes(bytes.try_into().unwrap());
-    NodeId::new(raw_id).map_err(|e| {
-        Error::Storage(StorageError::CorruptedData(format!(
-            "Invalid node ID in WAL {}: {}",
-            context, e
-        )))
-    })
-}
-
-/// Helper to deserialize and validate an EdgeId from WAL buffer
-#[inline]
-fn deserialize_edge_id(buffer: &[u8], offset: usize, context: &str) -> Result<EdgeId> {
-    let bytes = buffer.get(offset..offset + 8).ok_or_else(|| {
-        Error::Storage(StorageError::CorruptedData(format!(
-            "Insufficient buffer size for EdgeId in {}",
-            context
-        )))
-    })?;
-    let raw_id = u64::from_le_bytes(bytes.try_into().unwrap());
-    EdgeId::new(raw_id).map_err(|e| {
-        Error::Storage(StorageError::CorruptedData(format!(
-            "Invalid edge ID in WAL {}: {}",
-            context, e
-        )))
-    })
-}
-
-/// Helper to deserialize and validate a VersionId from WAL buffer
-#[inline]
-fn deserialize_version_id(buffer: &[u8], offset: usize, context: &str) -> Result<VersionId> {
-    let bytes = buffer.get(offset..offset + 8).ok_or_else(|| {
-        Error::Storage(StorageError::CorruptedData(format!(
-            "Insufficient buffer size for VersionId in {}",
-            context
-        )))
-    })?;
-    let raw_id = u64::from_le_bytes(bytes.try_into().unwrap());
-    VersionId::new(raw_id).map_err(|e| {
-        Error::Storage(StorageError::CorruptedData(format!(
-            "Invalid version ID in WAL {}: {}",
-            context, e
-        )))
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::core::interning::GLOBAL_INTERNER;
-    use crate::core::temporal::time;
-    use crate::storage::wal::serialization::serialize_entry_into;
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_read_empty_directory() {
-        let dir = TempDir::new().unwrap();
-        let entries = read_entries_from_dir(dir.path(), LSN(1)).unwrap();
-        assert!(entries.is_empty());
-    }
-
-    #[test]
-    fn test_read_nonexistent_segment() {
-        let dir = TempDir::new().unwrap();
-        let path = dir.path().join("nonexistent.log");
-        let entries = read_segment(&path, LSN(1)).unwrap();
-        assert!(entries.is_empty());
-    }
-
-    // =============================================================================
-    // TDD Tests for parse_entry_at() - Issue #218
-    // =============================================================================
-
-    #[test]
-    fn test_parse_entry_at_create_node() {
-        // Create a CreateNode entry
-        let node_id = NodeId::new(42).unwrap();
-        let operation = WalOperation::CreateNode {
-            node_id,
-            label: GLOBAL_INTERNER.intern("Person").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(1), operation);
-
-        // Serialize it
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        // Parse it back
-        let (parsed_entry, bytes_consumed) = parse_entry_at(&buffer, 0, WAL_VERSION).unwrap();
-
-        // Verify
-        assert_eq!(parsed_entry.lsn, LSN(1));
-        assert_eq!(bytes_consumed, buffer.len());
-        match parsed_entry.operation {
-            WalOperation::CreateNode {
-                node_id: parsed_id,
-                label,
-                ..
-            } => {
-                assert_eq!(parsed_id, node_id);
-                assert_eq!(label, GLOBAL_INTERNER.intern("Person").unwrap());
-            }
-            _ => panic!("Expected CreateNode operation"),
-        }
-    }
-
-    #[test]
-    fn test_parse_entry_at_create_edge() {
-        // Create a CreateEdge entry
-        let edge_id = EdgeId::new(100).unwrap();
-        let source = NodeId::new(1).unwrap();
-        let target = NodeId::new(2).unwrap();
-        let operation = WalOperation::CreateEdge {
-            edge_id,
-            source,
-            target,
-            label: GLOBAL_INTERNER.intern("KNOWS").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(2), operation);
-
-        // Serialize it
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        // Parse it back
-        let (parsed_entry, bytes_consumed) = parse_entry_at(&buffer, 0, WAL_VERSION).unwrap();
-
-        // Verify
-        assert_eq!(parsed_entry.lsn, LSN(2));
-        assert_eq!(bytes_consumed, buffer.len());
-        match parsed_entry.operation {
-            WalOperation::CreateEdge {
-                edge_id: parsed_id,
-                source: parsed_source,
-                target: parsed_target,
-                label,
-                ..
-            } => {
-                assert_eq!(parsed_id, edge_id);
-                assert_eq!(parsed_source, source);
-                assert_eq!(parsed_target, target);
-                assert_eq!(label, GLOBAL_INTERNER.intern("KNOWS").unwrap());
-            }
-            _ => panic!("Expected CreateEdge operation"),
-        }
-    }
-
-    #[test]
-    fn test_parse_entry_at_update_node() {
-        // Create an UpdateNode entry
-        let node_id = NodeId::new(42).unwrap();
-        let version_id = VersionId::new(1).unwrap();
-        let operation = WalOperation::UpdateNode {
-            node_id,
-            version_id,
-            label: GLOBAL_INTERNER.intern("UpdatedPerson").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(3), operation);
-
-        // Serialize it
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        // Parse it back
-        let (parsed_entry, bytes_consumed) = parse_entry_at(&buffer, 0, WAL_VERSION).unwrap();
-
-        // Verify
-        assert_eq!(parsed_entry.lsn, LSN(3));
-        assert_eq!(bytes_consumed, buffer.len());
-        match parsed_entry.operation {
-            WalOperation::UpdateNode {
-                node_id: parsed_id,
-                version_id: parsed_version,
-                label,
-                ..
-            } => {
-                assert_eq!(parsed_id, node_id);
-                assert_eq!(parsed_version, version_id);
-                assert_eq!(label, GLOBAL_INTERNER.intern("UpdatedPerson").unwrap());
-            }
-            _ => panic!("Expected UpdateNode operation"),
-        }
-    }
-
-    #[test]
-    fn test_parse_entry_at_update_edge() {
-        // Create an UpdateEdge entry
-        let edge_id = EdgeId::new(100).unwrap();
-        let version_id = VersionId::new(1).unwrap();
-        let operation = WalOperation::UpdateEdge {
-            edge_id,
-            version_id,
-            label: GLOBAL_INTERNER.intern("UPDATED_KNOWS").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(4), operation);
-
-        // Serialize it
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        // Parse it back
-        let (parsed_entry, bytes_consumed) = parse_entry_at(&buffer, 0, WAL_VERSION).unwrap();
-
-        // Verify
-        assert_eq!(parsed_entry.lsn, LSN(4));
-        assert_eq!(bytes_consumed, buffer.len());
-        match parsed_entry.operation {
-            WalOperation::UpdateEdge {
-                edge_id: parsed_id,
-                version_id: parsed_version,
-                label,
-                ..
-            } => {
-                assert_eq!(parsed_id, edge_id);
-                assert_eq!(parsed_version, version_id);
-                assert_eq!(label, GLOBAL_INTERNER.intern("UPDATED_KNOWS").unwrap());
-            }
-            _ => panic!("Expected UpdateEdge operation"),
-        }
-    }
-
-    #[test]
-    fn test_parse_entry_at_delete_node() {
-        // Create a DeleteNode entry
-        let node_id = NodeId::new(42).unwrap();
-        let operation = WalOperation::DeleteNode {
-            node_id,
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(5), operation);
-
-        // Serialize it
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        // Parse it back
-        let (parsed_entry, bytes_consumed) = parse_entry_at(&buffer, 0, WAL_VERSION).unwrap();
-
-        // Verify
-        assert_eq!(parsed_entry.lsn, LSN(5));
-        assert_eq!(bytes_consumed, buffer.len());
-        match parsed_entry.operation {
-            WalOperation::DeleteNode {
-                node_id: parsed_id, ..
-            } => {
-                assert_eq!(parsed_id, node_id);
-            }
-            _ => panic!("Expected DeleteNode operation"),
-        }
-    }
-
-    #[test]
-    fn test_parse_entry_at_delete_edge() {
-        // Create a DeleteEdge entry
-        let edge_id = EdgeId::new(100).unwrap();
-        let operation = WalOperation::DeleteEdge {
-            edge_id,
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(6), operation);
-
-        // Serialize it
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        // Parse it back
-        let (parsed_entry, bytes_consumed) = parse_entry_at(&buffer, 0, WAL_VERSION).unwrap();
-
-        // Verify
-        assert_eq!(parsed_entry.lsn, LSN(6));
-        assert_eq!(bytes_consumed, buffer.len());
-        match parsed_entry.operation {
-            WalOperation::DeleteEdge {
-                edge_id: parsed_id, ..
-            } => {
-                assert_eq!(parsed_id, edge_id);
-            }
-            _ => panic!("Expected DeleteEdge operation"),
-        }
-    }
-
-    #[test]
-    fn test_parse_entry_at_checkpoint() {
-        // Create a Checkpoint entry
-        let cp_timestamp = time::now();
-        let operation = WalOperation::Checkpoint {
-            lsn: LSN(100),
-            timestamp: cp_timestamp,
-        };
-        let entry = WalEntry::new(LSN(7), operation);
-
-        // Serialize it
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        // Parse it back
-        let (parsed_entry, bytes_consumed) = parse_entry_at(&buffer, 0, WAL_VERSION).unwrap();
-
-        // Verify
-        assert_eq!(parsed_entry.lsn, LSN(7));
-        assert_eq!(bytes_consumed, buffer.len());
-        match parsed_entry.operation {
-            WalOperation::Checkpoint { lsn, .. } => {
-                assert_eq!(lsn, LSN(100));
-            }
-            _ => panic!("Expected Checkpoint operation"),
-        }
-    }
-
-    #[test]
-    fn test_parse_entry_at_with_offset() {
-        // Create two entries
-        let operation1 = WalOperation::CreateNode {
-            node_id: NodeId::new(1).unwrap(),
-            label: GLOBAL_INTERNER.intern("First").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry1 = WalEntry::new(LSN(1), operation1);
-
-        let operation2 = WalOperation::CreateNode {
-            node_id: NodeId::new(2).unwrap(),
-            label: GLOBAL_INTERNER.intern("Second").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry2 = WalEntry::new(LSN(2), operation2);
-
-        // Serialize both entries separately, then concatenate
-        // (serialize_entry_into computes checksum from buffer start, so we can't
-        //  append directly without getting wrong checksums)
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry1, &mut buffer).unwrap();
-        let offset1_end = buffer.len();
-
-        let mut buffer2 = Vec::new();
-        serialize_entry_into(&entry2, &mut buffer2).unwrap();
-        buffer.extend_from_slice(&buffer2);
-
-        // Parse second entry using offset
-        let (parsed_entry, bytes_consumed) =
-            parse_entry_at(&buffer, offset1_end, WAL_VERSION).unwrap();
-
-        // Verify
-        assert_eq!(parsed_entry.lsn, LSN(2));
-        match parsed_entry.operation {
-            WalOperation::CreateNode { label, .. } => {
-                assert_eq!(label, GLOBAL_INTERNER.intern("Second").unwrap());
-            }
-            _ => panic!("Expected CreateNode operation"),
-        }
-        assert_eq!(bytes_consumed, buffer.len() - offset1_end);
-    }
-
-    #[test]
-    fn test_parse_entry_at_insufficient_buffer() {
-        // Create a buffer with only 10 bytes (not enough for LSN + timestamp + checksum)
-        let buffer = vec![0u8; 10];
-
-        // Should return error
-        let result = parse_entry_at(&buffer, 0, WAL_VERSION);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_parse_entry_at_unknown_operation_type() {
-        // Create a valid header but invalid operation type
-        let mut buffer = Vec::new();
-
-        // LSN (8 bytes)
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-
-        // Timestamp (12 bytes)
-        let timestamp = time::now();
-        timestamp.serialize_into(&mut buffer);
-
-        // Checksum (4 bytes) - just use 0 for this test
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // Invalid operation type (255)
-        buffer.push(255);
-
-        // Should return error for unknown operation type
-        let result = parse_entry_at(&buffer, 0, WAL_VERSION);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_parse_entry_at_truncated_operation_data() {
-        // Create a valid header but truncate operation data
-        let mut buffer = Vec::new();
-
-        // LSN (8 bytes)
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-
-        // Timestamp (12 bytes)
-        let timestamp = time::now();
-        timestamp.serialize_into(&mut buffer);
-
-        // Checksum (4 bytes)
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // Operation type for CreateNode (1)
-        buffer.push(1);
-
-        // Only 4 bytes of node_id (should be 8) - truncated!
-        buffer.extend_from_slice(&[1, 2, 3, 4]);
-
-        // Should return error for insufficient data
-        let result = parse_entry_at(&buffer, 0, WAL_VERSION);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_parse_entry_at_version_0_compatibility() {
-        // Test legacy version 0 parsing (without properties and temporal data)
-        // This tests the version < WAL_VERSION code path
-        let mut buffer = Vec::new();
-
-        // LSN (8 bytes)
-        buffer.extend_from_slice(&42u64.to_le_bytes());
-
-        // Timestamp (12 bytes)
-        let timestamp = time::now();
-        timestamp.serialize_into(&mut buffer);
-
-        // Placeholder checksum (4 bytes) - will be computed later
-        let checksum_offset = buffer.len();
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // Operation type: CreateNode (1)
-        buffer.push(1);
-
-        // Node ID (8 bytes)
-        buffer.extend_from_slice(&123u64.to_le_bytes());
-
-        // Label (4-byte InternedString ID)
-        let label_id = GLOBAL_INTERNER.intern("TestNode").unwrap().as_u32();
-        buffer.extend_from_slice(&label_id.to_le_bytes());
-
-        // Note: Version 0 format does NOT include properties or temporal data
-
-        // Compute checksum
-        let mut hasher = crc32fast::Hasher::new();
-        hasher.update(&buffer[0..checksum_offset]); // LSN + timestamp
-        hasher.update(&buffer[checksum_offset + 4..]); // Operation data
-        let checksum = hasher.finalize();
-        buffer[checksum_offset..checksum_offset + 4].copy_from_slice(&checksum.to_le_bytes());
-
-        // Parse with version 0
-        let (parsed_entry, bytes_consumed) = parse_entry_at(&buffer, 0, 0).unwrap();
-
-        // Verify
-        assert_eq!(parsed_entry.lsn.0, 42);
-        assert_eq!(bytes_consumed, buffer.len());
-        match parsed_entry.operation {
-            WalOperation::CreateNode {
-                node_id,
-                label: parsed_label,
-                properties,
-                valid_from,
-            } => {
-                assert_eq!(node_id.as_u64(), 123);
-                assert_eq!(parsed_label, GLOBAL_INTERNER.intern("TestNode").unwrap());
-                // Version 0 should have empty properties
-                assert!(properties.is_empty());
-                // Valid_from should be set to the timestamp
-                assert_eq!(valid_from, timestamp);
-            }
-            _ => panic!("Expected CreateNode operation"),
-        }
-    }
-
-    #[test]
-    fn test_parse_entry_at_checksum_mismatch() {
-        // Create a valid entry
-        let node_id = NodeId::new(42).unwrap();
-        let operation = WalOperation::CreateNode {
-            node_id,
-            label: GLOBAL_INTERNER.intern("Person").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(1), operation);
-
-        // Serialize it
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-
-        // Corrupt the checksum (bytes 20-24)
-        buffer[20] ^= 0xFF; // Flip all bits in first checksum byte
-
-        // Should return error for checksum mismatch
-        let result = parse_entry_at(&buffer, 0, WAL_VERSION);
-        assert!(result.is_err());
-        if let Err(e) = result {
-            let error_msg = format!("{}", e);
-            assert!(error_msg.contains("checksum mismatch"));
-        }
-    }
-
-    #[test]
-    fn test_parse_entry_at_update_edge_truncated_label() {
-        // Reproduction test for fuzzing panic: UpdateEdge with missing label
-        let mut buffer = Vec::new();
-
-        // LSN (8 bytes)
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-
-        // Timestamp (12 bytes)
-        let timestamp = time::now();
-        timestamp.serialize_into(&mut buffer);
-
-        // Checksum (4 bytes) - placeholders
-        let checksum_offset = buffer.len();
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // Operation type: UpdateEdge (4)
-        buffer.push(4);
-
-        // Edge ID (8 bytes)
-        buffer.extend_from_slice(&100u64.to_le_bytes());
-
-        // Version ID (8 bytes)
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-
-        // STOP HERE - Do not write label ID. This simulates truncation.
-        // We have written 16 bytes of operation data (EdgeID + VersionID), which satisfies the initial check.
-        // But we are missing the Label ID (4 bytes) which is read immediately after.
-
-        // Compute checksum for what we have
-        let mut hasher = crc32fast::Hasher::new();
-        hasher.update(&buffer[0..checksum_offset]); // LSN + timestamp
-        hasher.update(&buffer[checksum_offset + 4..]); // Operation data
-        let checksum = hasher.finalize();
-        buffer[checksum_offset..checksum_offset + 4].copy_from_slice(&checksum.to_le_bytes());
-
-        // Parse - this should NOT panic, but return an error
-        let result = parse_entry_at(&buffer, 0, WAL_VERSION);
-        assert!(result.is_err());
-
-        let err = result.unwrap_err();
-        let err_msg = format!("{}", err);
-        assert!(err_msg.contains("Insufficient buffer size"));
-    }
-
-    #[test]
-    fn test_parse_entry_at_update_node_truncated_label() {
-        // Reproduction test for fuzzing panic: UpdateNode with missing label
-        let mut buffer = Vec::new();
-
-        // LSN (8 bytes)
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-
-        // Timestamp (12 bytes)
-        let timestamp = time::now();
-        timestamp.serialize_into(&mut buffer);
-
-        // Checksum (4 bytes) - placeholders
-        let checksum_offset = buffer.len();
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // Operation type: UpdateNode (3)
-        buffer.push(3);
-
-        // Node ID (8 bytes)
-        buffer.extend_from_slice(&100u64.to_le_bytes());
-
-        // Version ID (8 bytes)
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-
-        // STOP HERE - Do not write label ID.
-
-        // Compute checksum for what we have
-        let mut hasher = crc32fast::Hasher::new();
-        hasher.update(&buffer[0..checksum_offset]); // LSN + timestamp
-        hasher.update(&buffer[checksum_offset + 4..]); // Operation data
-        let checksum = hasher.finalize();
-        buffer[checksum_offset..checksum_offset + 4].copy_from_slice(&checksum.to_le_bytes());
-
-        // Parse - this should NOT panic, but return an error
-        let result = parse_entry_at(&buffer, 0, WAL_VERSION);
-        assert!(result.is_err());
-
-        let err = result.unwrap_err();
-        let err_msg = format!("{}", err);
-        assert!(err_msg.contains("Insufficient buffer size"));
-    }
-
-    // =============================================================================
-    // TDD Tests for Memory-Efficient Segment Reading - Issue #216
-    // =============================================================================
-
-    /// Test that we can read a segment file with many entries without loading
-    /// the entire file into memory at once.
-    ///
-    /// This test creates a large segment file (simulating real-world 64MB segments)
-    /// and verifies that all entries can be read correctly.
-    #[test]
-    fn test_read_large_segment_memory_efficient() {
-        use std::io::Write;
-
-        let dir = TempDir::new().unwrap();
-        let segment_path = dir.path().join("large_segment.log");
-
-        // Create a segment file with many entries
-        let mut file = File::create(&segment_path).unwrap();
-
-        // Write WAL header
-        file.write_all(&WAL_MAGIC).unwrap();
-        file.write_all(&[WAL_VERSION]).unwrap();
-
-        // Create and write many entries to simulate a large segment
-        // We'll create 1000 entries, which should be several MB
-        let num_entries = 1000;
-        let mut expected_lsns = Vec::new();
-
-        for i in 0..num_entries {
-            let lsn = LSN(i + 1);
-            expected_lsns.push(lsn);
-
-            let operation = WalOperation::CreateNode {
-                node_id: NodeId::new(i + 1).unwrap(),
-                label: GLOBAL_INTERNER.intern(format!("Node_{}", i)).unwrap(),
-                properties: PropertyMap::new(),
-                valid_from: time::now(),
-            };
-
-            let entry = WalEntry::new(lsn, operation);
-            let mut buffer = Vec::new();
-            serialize_entry_into(&entry, &mut buffer).unwrap();
-            file.write_all(&buffer).unwrap();
-        }
-
-        file.sync_all().unwrap();
-        drop(file);
-
-        // Read the segment
-        let entries = read_segment(&segment_path, LSN(1)).unwrap();
-
-        // Verify all entries were read correctly
-        assert_eq!(entries.len(), num_entries as usize);
-        for (i, entry) in entries.iter().enumerate() {
-            assert_eq!(entry.lsn, LSN(i as u64 + 1));
-        }
-    }
-
-    /// Test that reading multiple segments doesn't accumulate excessive memory.
-    ///
-    /// This test creates multiple segment files and verifies that we can process
-    /// them sequentially without holding all segment buffers in memory simultaneously.
-    #[test]
-    fn test_read_multiple_segments_sequentially() {
-        use std::io::Write;
-
-        let dir = TempDir::new().unwrap();
-
-        // Create 5 segment files
-        let num_segments = 5;
-        let entries_per_segment = 100;
-
-        for seg_id in 0..num_segments {
-            let segment_path = dir.path().join(format!("{}.log", seg_id));
-            let mut file = File::create(&segment_path).unwrap();
-
-            // Write WAL header
-            file.write_all(&WAL_MAGIC).unwrap();
-            file.write_all(&[WAL_VERSION]).unwrap();
-
-            // Write entries for this segment
-            for i in 0..entries_per_segment {
-                let lsn = LSN((seg_id * entries_per_segment) + i + 1);
-
-                let operation = WalOperation::CreateNode {
-                    node_id: NodeId::new(lsn.0).unwrap(),
-                    label: GLOBAL_INTERNER
-                        .intern(format!("Node_seg{}_entry{}", seg_id, i))
-                        .unwrap(),
-                    properties: PropertyMap::new(),
-                    valid_from: time::now(),
-                };
-
-                let entry = WalEntry::new(lsn, operation);
-                let mut buffer = Vec::new();
-                serialize_entry_into(&entry, &mut buffer).unwrap();
-                file.write_all(&buffer).unwrap();
-            }
-
-            file.sync_all().unwrap();
-        }
-
-        // Read all entries from directory
-        let entries = read_entries_from_dir(dir.path(), LSN(1)).unwrap();
-
-        // Verify all entries were read correctly
-        assert_eq!(entries.len(), (num_segments * entries_per_segment) as usize);
-
-        // Verify entries are sorted by LSN
-        for i in 0..entries.len() - 1 {
-            assert!(entries[i].lsn <= entries[i + 1].lsn);
-        }
-    }
-
-    /// Test that segment reading works correctly with the start_lsn filter.
-    ///
-    /// This verifies that we can efficiently skip entries before a certain LSN
-    /// without processing them.
-    #[test]
-    fn test_read_segment_with_start_lsn_filter() {
-        use std::io::Write;
-
-        let dir = TempDir::new().unwrap();
-        let segment_path = dir.path().join("filtered_segment.log");
-
-        let mut file = File::create(&segment_path).unwrap();
-
-        // Write WAL header
-        file.write_all(&WAL_MAGIC).unwrap();
-        file.write_all(&[WAL_VERSION]).unwrap();
-
-        // Write 100 entries with LSN 1-100
-        for i in 1..=100 {
-            let lsn = LSN(i);
-            let operation = WalOperation::CreateNode {
-                node_id: NodeId::new(i).unwrap(),
-                label: GLOBAL_INTERNER.intern(format!("Node_{}", i)).unwrap(),
-                properties: PropertyMap::new(),
-                valid_from: time::now(),
-            };
-
-            let entry = WalEntry::new(lsn, operation);
-            let mut buffer = Vec::new();
-            serialize_entry_into(&entry, &mut buffer).unwrap();
-            file.write_all(&buffer).unwrap();
-        }
-
-        file.sync_all().unwrap();
-        drop(file);
-
-        // Read entries starting from LSN 50
-        let entries = read_segment(&segment_path, LSN(50)).unwrap();
-
-        // Should only get entries with LSN >= 50
-        assert_eq!(entries.len(), 51); // LSN 50-100 inclusive
-        assert_eq!(entries[0].lsn, LSN(50));
-        assert_eq!(entries[entries.len() - 1].lsn, LSN(100));
-    }
-
-    /// Test that empty segments are handled efficiently.
-    #[test]
-    fn test_read_empty_segment_efficient() {
-        use std::io::Write;
-
-        let dir = TempDir::new().unwrap();
-        let segment_path = dir.path().join("empty_segment.log");
-
-        let mut file = File::create(&segment_path).unwrap();
-
-        // Write only WAL header, no entries
-        file.write_all(&WAL_MAGIC).unwrap();
-        file.write_all(&[WAL_VERSION]).unwrap();
-
-        file.sync_all().unwrap();
-        drop(file);
-
-        // Read the empty segment
-        let entries = read_segment(&segment_path, LSN(1)).unwrap();
-
-        // Should return empty vector
-        assert!(entries.is_empty());
-    }
-
-    /// Test that partial/truncated entries at end of segment are handled gracefully.
-    ///
-    /// This can happen if a write was interrupted mid-entry.
-    #[test]
-    fn test_read_segment_with_truncated_entry() {
-        use std::io::Write;
-
-        let dir = TempDir::new().unwrap();
-        let segment_path = dir.path().join("truncated_segment.log");
-
-        let mut file = File::create(&segment_path).unwrap();
-
-        // Write WAL header
-        file.write_all(&WAL_MAGIC).unwrap();
-        file.write_all(&[WAL_VERSION]).unwrap();
-
-        // Write one complete entry
-        let operation = WalOperation::CreateNode {
-            node_id: NodeId::new(1).unwrap(),
-            label: GLOBAL_INTERNER.intern("Node_1").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(1), operation);
-        let mut buffer = Vec::new();
-        serialize_entry_into(&entry, &mut buffer).unwrap();
-        file.write_all(&buffer).unwrap();
-
-        // Write a partial entry (just the LSN, incomplete)
-        file.write_all(&42u64.to_le_bytes()).unwrap();
-
-        file.sync_all().unwrap();
-        drop(file);
-
-        // Read the segment - should get the complete entry and stop at truncation
-        let entries = read_segment(&segment_path, LSN(1)).unwrap();
-
-        // Should only get the one complete entry
-        assert_eq!(entries.len(), 1);
-        assert_eq!(entries[0].lsn, LSN(1));
-    }
-
-    // =============================================================================
-    // Security and Error Handling Tests - Issue #216 Fixes
-    // =============================================================================
-
-    /// Test that non-existent files return empty results (not an error).
-    #[test]
-    fn test_read_nonexistent_file_returns_empty() {
-        let dir = TempDir::new().unwrap();
-        let nonexistent = dir.path().join("does_not_exist.log");
-
-        // Should return Ok(empty vector), not an error
-        let result = read_segment(&nonexistent, LSN(1));
-        assert!(result.is_ok());
-        assert!(result.unwrap().is_empty());
-    }
-
-    /// Test that file size validation prevents reading excessively large files.
-    ///
-    /// This protects against DoS attacks where an attacker places a huge file
-    /// in the WAL directory.
-    #[test]
-    fn test_read_segment_rejects_oversized_file() {
-        use std::io::Write;
-
-        let dir = TempDir::new().unwrap();
-        let segment_path = dir.path().join("oversized_segment.log");
-
-        let mut file = File::create(&segment_path).unwrap();
-
-        // Write WAL header
-        file.write_all(&WAL_MAGIC).unwrap();
-        file.write_all(&[WAL_VERSION]).unwrap();
-
-        // Seek to a position beyond MAX_SEGMENT_SIZE (1GB)
-        // Note: We don't actually write 1GB of data, just seek past it
-        // This creates a sparse file that reports a large size
-        const OVERSIZED: u64 = 1024 * 1024 * 1024 + 1; // 1GB + 1 byte
-        file.set_len(OVERSIZED).unwrap();
-
-        file.sync_all().unwrap();
-        drop(file);
-
-        // Should return an error about file being too large
-        let result = read_segment(&segment_path, LSN(1));
-        assert!(result.is_err());
-        let error_msg = format!("{}", result.unwrap_err());
-        assert!(
-            error_msg.contains("too large"),
-            "Expected 'too large' error, got: {}",
-            error_msg
-        );
-    }
-
-    #[test]
-    fn test_wal_offset_overflow_protection() {
-        // Create a small dummy buffer
-        let buffer = [0u8; 100];
-
-        // Use an offset close to usize::MAX
-        let offset = usize::MAX - 10;
-
-        // Attempt to parse - this should trigger the checked_add protection
-        // NOT a panic or buffer overrun
-        let result = parse_entry_at(&buffer, offset, 1);
-
-        assert!(result.is_err());
-        match result {
-            Err(Error::Storage(StorageError::CorruptedData(msg))) => {
-                assert_eq!(msg, "WAL offset overflow");
-            }
-            _ => panic!("Expected WAL offset overflow error, got: {:?}", result),
-        }
-    }
-
-    #[test]
-    fn test_update_node_insufficient_buffer_for_label() {
-        // Create a valid UpdateNode entry
-        let node_id = NodeId::new(42).unwrap();
-        let version_id = VersionId::new(1).unwrap();
-        let operation = WalOperation::UpdateNode {
-            node_id,
-            version_id,
-            label: GLOBAL_INTERNER.intern("UpdatedPerson").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(1), operation);
-
-        // Serialize it
-        let mut full_buffer = Vec::new();
-        serialize_entry_into(&entry, &mut full_buffer).unwrap();
-
-        // Calculate expected cut point
-        // Header (24) + Op (1) + NodeID (8) + VersionID (8) = 41 bytes
-        // We want to pass the first check (41 bytes) but fail the next (Label ID, +4 bytes)
-        // So we truncate to EXACTLY 41 bytes.
-        let truncated_buffer = &full_buffer[0..41];
-
-        // This should trigger "Insufficient buffer size for UpdateNode label"
-        let result = parse_entry_at(truncated_buffer, 0, WAL_VERSION);
-        assert!(result.is_err());
-        if let Err(Error::Storage(StorageError::CorruptedData(msg))) = result {
-            assert_eq!(msg, "Insufficient buffer size for UpdateNode label");
-        } else {
-            panic!("Expected specific CorruptedData error, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_update_edge_insufficient_buffer_for_label() {
-        // Create a valid UpdateEdge entry
-        let edge_id = EdgeId::new(100).unwrap();
-        let version_id = VersionId::new(1).unwrap();
-        let operation = WalOperation::UpdateEdge {
-            edge_id,
-            version_id,
-            label: GLOBAL_INTERNER.intern("UPDATED_KNOWS").unwrap(),
-            properties: PropertyMap::new(),
-            valid_from: time::now(),
-        };
-        let entry = WalEntry::new(LSN(1), operation);
-
-        // Serialize it
-        let mut full_buffer = Vec::new();
-        serialize_entry_into(&entry, &mut full_buffer).unwrap();
-
-        // Calculate expected cut point.
-        // UpdateEdge now validates all V1 fixed fields in one check:
-        // Header (24) + Op (1) + EdgeID (8) + VersionID (8) + LabelID (4) = 45 bytes.
-        // Truncating to 41 bytes should fail the fixed-fields boundary check.
-        let truncated_buffer = &full_buffer[0..41];
-
-        // This should trigger the generic UpdateEdge insufficient buffer error.
-        let result = parse_entry_at(truncated_buffer, 0, WAL_VERSION);
-        assert!(result.is_err());
-        if let Err(Error::Storage(StorageError::CorruptedData(msg))) = result {
-            assert_eq!(msg, "Insufficient buffer size for UpdateEdge");
-        } else {
-            panic!("Expected specific CorruptedData error, got: {:?}", result);
-        }
-    }
-
-    #[test]
-    fn test_update_edge_offset_overflow_before_label() {
-        // This test attempts to trigger the overflow check before reading the label ID in UpdateEdge
-        // It's hard to trigger purely via buffer offset manipulation without triggering earlier checks,
-        // unless we mock the buffer length check or construct a very specific scenario.
-        //
-        // However, we can construct a buffer that passes earlier checks but fails the overflow check
-        // if we use a huge offset that wraps around when adding 4.
-        //
-        // Let's try to pass a buffer and an offset such that offset + 16 (for edge+ver) succeeds,
-        // but offset + 16 + 4 overflows.
-        //
-        // offset + 16 <= usize::MAX
-        // offset + 20 > usize::MAX (overflow)
-        // So offset can be usize::MAX - 19.
-
-        // We need a buffer that is technically "valid" up to that point logic-wise,
-        // but since we are passing a huge offset, we need the buffer length to be huge too?
-        // No, `buffer.len()` is checked against `current_offset`.
-        // If `current_offset` is huge, `buffer.len()` must be huge for the check `current_offset > buffer.len()` to pass.
-        // Since we can't allocate a usize::MAX buffer, we can't easily test the "success" path up to the overflow.
-        //
-        // BUT, the `checked_add` returns None on overflow, and we convert that to an error.
-        // So we just need `current_offset.checked_add(4)` to return None.
-        // And we need to get past the previous checks.
-        //
-        // Previous checks in UpdateEdge:
-        // 1. `current_offset.checked_add(16)` (Edge ID + Version ID)
-        //
-        // So if we start with an offset that allows +16 but fails +20 (implicit in logic flow),
-        // we might hit it. But `parse_entry_at` starts from `offset`.
-        //
-        // The function does:
-        // header checks (offset + 24) -> OK
-        // op type check (offset + 1) -> OK
-        // UpdateEdge checks:
-        //   offset + 16 -> OK
-        //   read edge_id, version_id -> OK
-        //   offset + 4 -> OVERFLOW?
-        //
-        // To get to UpdateEdge check, we need to pass header checks.
-        // `offset + 24` must not overflow.
-        // So `offset` must be <= usize::MAX - 24.
-        //
-        // Inside UpdateEdge:
-        // `current_offset` is now `offset + 24 + 1` (header + op type) = `offset + 25`.
-        // Then checks `current_offset + 16`. `offset + 25 + 16` = `offset + 41`.
-        // Then adds 16. `current_offset` is `offset + 41`.
-        // Then checks `current_offset + 4`. `offset + 41 + 4` = `offset + 45`.
-        //
-        // So if we pick `offset` such that `offset + 45` overflows, but `offset + 41` does not?
-        // Yes. `usize::MAX - 44`.
-        // `offset + 41` = `MAX - 3` (OK)
-        // `offset + 45` = OVERFLOW (Error)
-        //
-        // However, we also need `current_offset < buffer.len()`.
-        // `buffer.len()` would need to be `usize::MAX - 3`. We can't allocate that.
-        //
-        // So we can't integration-test the overflow check with a real buffer on a 64-bit machine.
-        // But on a 32-bit machine (or if we could mock the buffer), maybe.
-        //
-        // Actually, the `checked_add` protection is `ok_or_else(|| Error...)`.
-        // This error `WAL offset overflow` is what we want to verify.
-        //
-        // Since we can't allocate a huge buffer, this test is theoretical unless we can mock `buffer.len()` or use a trick.
-        // The check is `checked_add(...) > buffer.len()`.
-        // If `checked_add` fails (returns None), we get the error immediately.
-        // We don't check buffer length if `checked_add` fails.
-        //
-        // So if we pass a small buffer, but a huge offset?
-        // Then `current_offset > buffer.len()` check inside `add_offset!` or manual checks will fail
-        // with "Insufficient buffer size..." BEFORE we get to the overflow check?
-        //
-        // Let's trace:
-        // `parse_entry_at(buffer, offset)`
-        // `current_offset = offset`
-        // `if current_offset.checked_add(24)... > buffer.len()` -> Error "Insufficient buffer size..."
-        //
-        // So we can never get past the first check with a huge offset and a small buffer.
-        // Thus, we can't easily test the later overflow checks without a huge buffer.
-        //
-        // Use `#[cfg(target_pointer_width = "32")]`? No, CI is likely 64-bit.
-        //
-        // However, the coverage report says lines 518-520 are missed.
-        // `src/storage/wal/segment_reader.rs:518`:
-        // if current_offset.checked_add(4).ok_or_else(|| ...
-        //
-        // Wait, if I can't reach it, maybe it's dead code?
-        // No, it's valid protection.
-        //
-        // Actually, the previous test `test_wal_offset_overflow_protection` just calls `parse_entry_at` with huge offset.
-        // And it hits the FIRST check: `checked_add(24)`.
-        //
-        // To hit the UpdateEdge specific overflow check, we'd need to pass the first check.
-        //
-        // What if we test the logic in isolation? We can't, it's inside the function.
-        //
-        // Let's settle for testing the `Insufficient buffer size` error, which IS reachable with small buffers.
-        // The overflow check is likely unreachable in tests without huge buffers, so we might have to accept it as uncovered or add `// LCOV_EXCL_START`?
-        // But the user wants coverage.
-        //
-        // Wait, Codecov says lines 518-520 are uncovered.
-        // Line 518 is the `if current_offset.checked_add(4)...` check.
-        //
-        // If I supply a buffer that is large enough to pass the *previous* checks but *truncated* right after,
-        // then `checked_add(4)` will succeed (return Some), but `> buffer.len()` will be true.
-        // This will verify the logic `> buffer.len()` branch.
-        //
-        // The `WAL offset overflow` error (from `.ok_or_else`) is what handles the arithmetic overflow.
-        // The `Insufficient buffer size` error is what handles the buffer boundary.
-        //
-        // My proposed `test_update_edge_insufficient_buffer_for_label` will cover the `Insufficient buffer size` path.
-        //
-        // Is line 518 the check itself? Yes.
-        // If the test runs, it executes the line `if current_offset.checked_add(4)...`.
-        // Even if it doesn't panic/return overflow error, it executes the condition.
-        //
-        // Codecov usually marks the line as covered if it's executed.
-        //
-        // So `test_update_edge_insufficient_buffer_for_label` should cover lines 518-520 (the condition) and 524 (the error return).
-        //
-        // The overflow branch (inside `ok_or_else`) might remain uncovered, but that's fine if the main path is covered.
-    }
-}
-
-#[cfg(test)]
-mod regression_tests {
-    use super::*;
-
-    #[test]
-    fn test_repro_fuzz_update_edge_panic() {
-        // Failing input from fuzzer:
-        // [71, 87, 65, 76, 1, 190, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 40, 1, 1, 1, 1, 1, 71, 87, 65, 76, 0, 4, 0, 0, 0, 1, 40, 1, 1, 1, 1, 1, 71, 87, 65, 76, 76, 0, 0, 0]
-        let data = vec![
-            71, 87, 65, 76, 1, // Header: GWAL, Ver 1
-            190, 0, 0, 0, 0, 0, 0, 0, // LSN: 190
-            0, 1, 1, 1, 1, 40, 1, 1, 1, 1, 1, 71, // Timestamp (12 bytes)
-            87, 65, 76, 0, // Checksum (4 bytes)
-            4, // OpType: 4 (UpdateEdge)
-            0, 0, 0, 1, 40, 1, 1, 1, // EdgeId (8 bytes)
-            1, 1, 71, 87, 65, 76, 76,
-            0, // VersionId (8 bytes)
-               // Total length: 48 bytes
-               // Missing LabelId (4 bytes) required for Ver 1
-        ];
-
-        // Offset 5 to skip header
-        let result = parse_entry_at(&data, 5, 1);
-
-        // Before fix: Panics with index out of bounds
-        // After fix: Returns Error
-        assert!(
-            result.is_err(),
-            "Should return error for truncated buffer, got {:?}",
-            result
-        );
-    }
-}
-
-#[cfg(test)]
-mod fuzz_tests {
-    use super::*;
-    use proptest::prelude::*;
-
-    proptest! {
-        // Fuzz parse_entry_at with arbitrary bytes
-        #[test]
-        fn fuzz_parse_entry_at(
-            bytes in prop::collection::vec(any::<u8>(), 0..2048),
-            offset in 0..100usize,
-            version in 0..2u8
-        ) {
-            // Should not panic
-            let _ = parse_entry_at(&bytes, offset, version);
-        }
-    }
-}
-
-#[cfg(test)]
-mod sentry_tests {
-    use super::*;
-    use std::fs::File;
-    use std::io::Write;
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_read_segment_exactly_max_size_allowed() {
-        // 🛡️ Sentry Test: Verify read_segment allows file of exactly MAX_SEGMENT_SIZE.
-        // This targets mutants that change `>` to `>=`.
-        let dir = TempDir::new().unwrap();
-        let segment_path = dir.path().join("max_size.log");
-
-        let mut file = File::create(&segment_path).unwrap();
-
-        // Write header
-        file.write_all(&WAL_MAGIC).unwrap();
-        file.write_all(&[WAL_VERSION]).unwrap();
-
-        // Seek to exact MAX_SEGMENT_SIZE (sparse file)
-        file.set_len(MAX_SEGMENT_SIZE).unwrap();
-
-        drop(file);
-
-        // Read segment - should succeed (return empty entries or corruption error, but NOT "too large").
-        let result = read_segment(&segment_path, LSN(1));
-
-        match result {
-            Ok(_) => {
-                // Success is fine (e.g. if sparse zeros are skipped or interpreted as empty)
-            }
-            Err(e) => {
-                let msg = e.to_string();
-                assert!(
-                    !msg.contains("too large"),
-                    "Should not reject max size file. Error was: {}",
-                    msg
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_read_segment_header_only() {
-        // 🛡️ Sentry Test: Verify read_segment handles file with ONLY header (5 bytes).
-        // This targets mutants that change `>=` to `>` in header size check.
-        let dir = TempDir::new().unwrap();
-        let segment_path = dir.path().join("header_only.log");
-
-        let mut file = File::create(&segment_path).unwrap();
-        file.write_all(&WAL_MAGIC).unwrap();
-        file.write_all(&[WAL_VERSION]).unwrap();
-        // Total size = 5 bytes.
-        drop(file);
-
-        let result = read_segment(&segment_path, LSN(1));
-
-        assert!(result.is_ok(), "Should accept header-only segment");
-        assert!(result.unwrap().is_empty());
-    }
-
-    #[test]
-    fn test_parse_entry_at_exact_header_size() {
-        // 🛡️ Sentry Test: Verify parse_entry_at behavior with exactly 24 bytes (header size).
-        // Targets `>` vs `>=` in `if current_offset.checked_add(24)? > buffer.len()`.
-
-        // 24 bytes buffer
-        let buffer = vec![0u8; 24];
-
-        let result = parse_entry_at(&buffer, 0, WAL_VERSION);
-
-        assert!(result.is_err());
-        let msg = result.unwrap_err().to_string();
-        // We expect it to pass the first check (24 !> 24) and fail the op-type check.
-        assert!(
-            msg.contains("operation type"),
-            "Should fail at op type check, not header check. Got: {}",
-            msg
-        );
-    }
-
-    #[test]
-    fn test_parse_entry_at_exact_header_and_op_type() {
-        // 🛡️ Sentry Test: Verify parse_entry_at behavior with exactly 25 bytes (header + op type).
-        // Targets `>` vs `>=` in `if current_offset >= buffer.len()`.
-
-        let mut buffer = vec![0u8; 25];
-        // LSN=0, TS=0, Checksum=0.
-        // OpType = 255 (Unknown) at index 24.
-        buffer[24] = 255;
-
-        let result = parse_entry_at(&buffer, 0, WAL_VERSION);
-
-        assert!(result.is_err());
-        let msg = result.unwrap_err().to_string();
-        assert!(
-            msg.contains("Unknown WAL operation type"),
-            "Should read op type and fail validation. Got: {}",
-            msg
-        );
-    }
-
-    #[test]
-    fn test_bolt_pre_allocate_segment_capacity() {
-        use std::io::Write;
-        let dir = tempfile::TempDir::new().unwrap();
-        let file_path = dir.path().join("1.log");
-        let mut file = std::fs::File::create(&file_path).unwrap();
-
-        // Write magic header and some dummy data to make buffer large enough
-        let mut buffer = Vec::new();
-        buffer.extend_from_slice(&super::WAL_MAGIC);
-        buffer.push(super::WAL_VERSION); // Version
-
-        // Add padding to make the file size larger (e.g., 1024 bytes)
-        buffer.extend(vec![0; 1024 - buffer.len()]);
-        file.write_all(&buffer).unwrap();
-        file.sync_all().unwrap();
-
-        let entries = read_segment(&file_path, crate::storage::LSN(1)).unwrap();
-
-        // 1024 / 128 = 8. Since we expect capacity_hint = buffer.len() / 128
-        assert!(
-            entries.capacity() >= 8,
-            "⚡ Bolt: Vector should be pre-allocated with capacity based on file size. Capacity was {}",
-            entries.capacity()
-        );
-    }
-}
